+++ src/main/java/com/mclods/store_app/domain/dtos/common/UserDto.java
@@ -1,0 +1,22 @@
+package com.mclods.store_app.domain.dtos.common;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserDto {
+    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
+    private Long id;
+
+    private String name;
+
+    private String email;
+
+    private String password;
+}
